// src/shared/types/types.ts
import dayjs from 'dayjs';
import { SorterResult, TablePaginationConfig } from 'antd/es/table/interface';
import { GetProp, TableProps } from 'antd';
import { OrderingMap, SORT_NAMES } from '../const/const';

export type TSortDirection = keyof typeof SORT_NAMES;

export type OrderingType = typeof OrderingMap[keyof typeof OrderingMap];

export interface PaginationResult<T> {
    objects: T[]
    total?: number
}

/**
 * Base entity fields generated by the server
 * Used as a base for all ReadSchema types
 */
export interface BaseEntityFields {
    uuid: string;
    created_at: string;
    updated_at: string;
    deleted_at?: string;
}

export interface TableParams {
    pagination?: TablePaginationConfig;
    sortField?: SorterResult<any>['field'];
    sortOrder?: SorterResult<any>['order'];
    filters?: Parameters<GetProp<TableProps, 'onChange'>>[1];
}

export type TDateRangePicker = [
    start: dayjs.Dayjs | null | undefined,
    end: dayjs.Dayjs | null | undefined,
];

export interface FileTypeS3 {
    uuid: string
    created_at?: string
    content_type?: string
    updated_at?: string
    bucket?: string
    s3_id?: string
    filename: string
    url?: string
}
export interface FileType {
    id: number
    file?: File
    url?: string
    type?: string
    name?: string
}

export type TableRowSelection<T extends object = object> =
    TableProps<T>['rowSelection'];